move expo level store
